fix:修改 数据类型

+++ ses-api/ses-api-mobile-b/src/main/java/com/redescooter/ses/api/mobile/b/vo/express/OrderResult.java
@@ -194,8 +194,8 @@ public class OrderResult extends GeneralResult {
     private Date createdTime;
 
 
-    @ApiModelProperty(value="更新人")
-    private Integer updatedBy;
+    @ApiModelProperty(value = "更新人")
+    private Long updatedBy;
 
 
     @ApiModelProperty(value="更新时间")
